+++ blocks/details/src/edit.js
@@ -121,7 +121,7 @@ export default function Edit( object ) {
 	return (
 		<div { ...useBlockProps() }>
 			<InspectorControls>
-				<PanelBody initialOpen={false} title={ __( 'Settings', 'personio-integration-light' ) }>
+				<PanelBody title={ __( 'Settings', 'personio-integration-light' ) }>
           <div className="wp-personio-integration-selectcontrol">
             {
               <SelectControl