+++ changelog.md
@@ -26,6 +26,7 @@
 - Added info in admin footer if a page from our plugin is loaded
 - Added option to import project configuration during setup
 - Added log for error 500 during imports which also prevents hanging import tasks
+- Added option to reset the plugin in backend settings (in preparation for Cyber Resilience Act)
 
 ### Changed
 
@@ -45,6 +46,8 @@
 - Renamed "Application link" to "Option to apply"
 - Renamed "Content" to "Description"
 - Show hint in backend if no description is available for single position
+- Block settings are now visible on load
+- All Blocks are now in their own category "Personio Integration"
 - The help will show primary the help for the actual called page in backend
 - Optimized handling of header and footer settings for our block templates on theme switch
 - Checking whether a required plugin for an extension is installed (without it having to be activated)